Works with java 1.7u5 (needed new version for xsom and aspectj, and some fixes in code)

+++ infra/prism/src/main/java/com/evolveum/midpoint/prism/schema/DomToSchemaProcessor.java
@@ -481,8 +481,8 @@ class DomToSchemaProcessor {
 				itemDef.setMaxOccurs(-1);
 			}
 		} else {
-			itemDef.setMinOccurs(particle.getMinOccurs());
-			itemDef.setMaxOccurs(particle.getMaxOccurs());
+			itemDef.setMinOccurs(particle.getMinOccurs().intValue());
+			itemDef.setMaxOccurs(particle.getMaxOccurs().intValue());
 		}
 	}
